SQLite Database Archive
Added a SQLite Database Archive of opened files, as well as a history
tab that shows the details stored in the database. This includes
indication of which services the files have been uploaded to as well as
links to open the file in default web browser. Also tweaked the
RideWithGPS API usage after receiving apikey and docs from RideWithGPS
authors (thanks) so it now recods the ride name and description for this
service.